Fix 1q, 2w, etc. combos with Default Layer
These combos always sent GUI+number key, but this prevented them for
working on the MacOS layer. It should send Option+number key instead.

This fixes this. It's a lazy implementation because my C knowledge is
not that good enough. And I don't have much time either.

It works for now.